+++ CHANGELOG.md
@@ -66,6 +66,8 @@ const firebaseConfig = {
 const { login } = useAuth(firebaseConfig)
 ```
 
+- Remove BDrawer, use BPanel instead
+
 ## v0.2.0
 
 - [ui] BModal can be closed with esc key